requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
cmake==3.20.2
colorama==0.4.4
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.20.3
Pillow==8.2.0
requests==2.25.1
SQLAlchemy==1.4.15
urllib3==1.26.4
Werkzeug==2.0.1
